🌍 #31 More ExampleSpace tidy ups
- Break up the big ExampleSpace file
- Clearly redefine exploration stage as a sum-type
